Split up routes file
