+++ components/fires-server/app/exceptions/handler.ts
@@ -30,7 +30,9 @@ export default class HttpExceptionHandler extends ExceptionHandler {
    * @note You should not attempt to send a response from this method.
    */
   async report(error: unknown, ctx: HttpContext) {
-    ctx.logger.error(error)
+    if (app.inDev) {
+      ctx.logger.error(error)
+    }
     return super.report(error, ctx)
   }
 }
